modules/analytics_reports.py
# modules/analytics_reports.py
# This module is a placeholder for the "Analytics & Reports" feature.
# It envisions providing insights into prompt usage, performance, and suggestions.

import streamlit as st
import logging
import pandas as pd
import numpy as np
import altair as alt # For conceptual data visualization
from modules import shared_utils # Import shared utility functions

logger = logging.getLogger(__name__)

# ====================================================================================================
# SECTION 1: MODULE-SPECIFIC CONSTANTS AND CONFIGURATIONS
# Defines conceptual data and help messages.
# ====================================================================================================

ANALYTICS_HELP_MESSAGE = """
### Gain Insights from Your Prompt Usage! (Coming Soon!)
This module will provide valuable analytics and reports on how your prompts are performing.
**Envisioned Features:**
* **Prompt usage trends:** Track generation frequency, saves, and shares.
* **Best performing prompts:** Identify prompts leading to high satisfaction.
* **AI model comparison chart:** (Conceptual) Compare different LLM performances.
* **Suggest new prompt templates:** AI-driven ideas based on usage patterns.
* **Weekly report via email:** Automated summaries of your prompt activity.
* **AI-generated prompt improvement suggestions:** Based on aggregate data.
* **Generate prompt documentation:** Auto-create documentation for your prompts.
* **Usage heatmap:** Visualize popular features or prompt types.

Optimize your prompt engineering workflow with data-driven decisions!
"""

# Conceptual Data for Simulation
# In a real app, this would come from a backend database tracking user interactions.
CONCEPTUAL_USAGE_DATA = {
    "date": pd.to_datetime(["2024-01-01", "2024-01-08", "2024-01-15", "2024-01-22", "2024-01-29",
                            "2024-02-05", "2024-02-12", "2024-02-19", "2024-02-26", "2024-03-04"]),
    "prompts_generated": [50, 65, 70, 85, 90, 75, 100, 110, 95, 120],
    "prompts_saved": [10, 15, 12, 18, 20, 15, 25, 28, 22, 30],
    "chat_interactions": [20, 25, 30, 35, 40, 30, 45, 50, 40, 55],
}
CONCEPTUAL_USAGE_DF = pd.DataFrame(CONCEPTUAL_USAGE_DATA)

CONCEPTUAL_PROMPT_PERFORMANCE = {
    "Prompt Name": ["Creative Story Starter", "Email for Client Update", "Python Function", "Social Media Post", "Debate Argument"],
    "Generations": [250, 300, 180, 400, 120],
    "Saves": [50, 70, 30, 90, 25],
    "Shares": [10, 5, 2, 20, 3],
    "Avg_Rating": [4.5, 4.2, 3.8, 4.7, 4.0] # Conceptual rating
}
CONCEPTUAL_PERFORMANCE_DF = pd.DataFrame(CONCEPTUAL_PROMPT_PERFORMANCE)

# ====================================================================================================
# SECTION 2: CONCEPTUAL ANALYTICS LOGIC
# These functions simulate data aggregation and insight generation.
# ====================================================================================================

def conceptual_generate_usage_trends_chart(df: pd.DataFrame):
    """
    (Conceptual) Generates an Altair chart for usage trends.
    """
    if df.empty:
        st.warning("No usage data available for charting.")
        return

    # Melt DataFrame for multi-line chart
    melted_df = df.melt('date', var_name='Metric', value_name='Count')

    chart = alt.Chart(melted_df).mark_line(point=True).encode(
        x=alt.X('date', title='Date', axis=alt.Axis(format="%Y-%m-%d")),
        y=alt.Y('Count', title='Number of Actions'),
        color=alt.Color('Metric', title='Action Type'),
        tooltip=['date', 'Metric', 'Count']
    ).properties(
        title='Conceptual PromptGPT Suite Usage Trends'
    ).interactive()
    st.altair_chart(chart, use_container_width=True)
    logger.info("Conceptual usage trends chart generated.")

def conceptual_generate_best_prompts_chart(df: pd.DataFrame):
    """
    (Conceptual) Generates an Altair chart for best performing prompts.
    """
    if df.empty:
        st.warning("No prompt performance data available for charting.")
        return

    # Sort by a conceptual "performance score" (e.g., based on Generations, Saves, Shares, Rating)
    df['Performance_Score'] = (df['Generations'] * 0.4) + (df['Saves'] * 0.3) + (df['Shares'] * 0.2) + (df['Avg_Rating'] * 10)
    df_sorted = df.sort_values('Performance_Score', ascending=False).head(5)

    chart = alt.Chart(df_sorted).mark_bar().encode(
        x=alt.X('Performance_Score', title='Performance Score (Conceptual)'),
        y=alt.Y('Prompt Name', sort='-x', title='Prompt Template'),
        tooltip=['Prompt Name', 'Generations', 'Saves', 'Shares', 'Avg_Rating', 'Performance_Score']
    ).properties(
        title='Top 5 Conceptual Performing Prompts'
    ).interactive()
    st.altair_chart(chart, use_container_width=True)
    logger.info("Conceptual best prompts chart generated.")

# ====================================================================================================
# SECTION 3: STREAMLIT UI LAYOUT AND INTERACTION
# This section defines the user interface for the Analytics & Reports module.
# ====================================================================================================

def run(api_key_valid: bool):
    """
    Main function to run the Analytics & Reports module's Streamlit UI.
    This function is called by app.py when the 'Analytics & Reports' module is selected.
    """
    st.header("📊 Analytics & Reports")
    st.markdown("Track usage, analyze performance, and gain insights into your prompt engineering.")
    st.markdown("---")

    shared_utils.display_module_help(ANALYTICS_HELP_MESSAGE)

    st.warning("This module is conceptual. The analytics and reporting features will be fully implemented in future updates with real data integration. Below are simulated charts and reports.")
    st.markdown("---")

    # 3.1 Conceptual Usage Trends
    st.subheader("📈 Prompt Usage Trends (Simulated)")
    st.write("Monitor how often prompts are generated, saved, and chat interactions occur over time.")
    conceptual_generate_usage_trends_chart(CONCEPTUAL_USAGE_DF)

    st.markdown("---")

    # 3.2 Conceptual Best Performing Prompts
    st.subheader("🏆 Best Performing Prompts (Simulated)")
    st.write("Identify which of your prompt templates are most frequently used, saved, and shared.")
    conceptual_generate_best_prompts_chart(CONCEPTUAL_PERFORMANCE_DF)

    st.markdown("---")

    # 3.3 Conceptual AI-Generated Suggestions
    st.subheader("✨ AI-Generated Improvement Suggestions (Conceptual)")
    st.write("Leverage AI to suggest ways to improve your prompt engineering based on simulated patterns.")
    if st.button("Generate Conceptual Suggestions", key="ar_generate_suggestions"):
        st.info("Analyzing simulated data... (This would involve an AI call in a real app)")
        conceptual_suggestion = (
            "Based on simulated user behavior, prompts with explicit 'Chain-of-thought' instructions "
            "show a higher retention rate. Consider integrating this strategy more often. "
            "Also, prompts related to 'coding' tasks are highly saved, suggesting a need for more code-related templates."
        )
        st.markdown(f"**Conceptual Insight:**\n{conceptual_suggestion}")
    else:
        st.info("Click the button to get conceptual AI-generated insights.")


    # 3.4 Conceptual Reporting Options
    st.subheader("📥 Reporting Options (Conceptual)")
    st.markdown("Generate and export various reports.")
    col_pdf, col_email = st.columns(2)
    with col_pdf:
        if st.button("Download PDF Report (Conceptual)", use_container_width=True, key="ar_download_pdf"):
            st.success("Conceptual PDF report generated for download (not actually downloading for demo).")
            st.info("In a full app, this would trigger a PDF download of all analytics.")
    with col_email:
        if st.button("Schedule Weekly Email Report (Conceptual)", use_container_width=True, key="ar_schedule_email"):
            st.success("Conceptual weekly email report scheduled! (Functionality not active in demo).")
            st.info("This would set up a recurring email with usage stats.")

    shared_utils.display_ai_powered_notice()
    logger.info("Analytics & Reports module UI rendered (conceptual).")
